fix(prefetch): update anchor links to remove `rel=prefetch`
After confirming that changing the `prefetch` integration selector to target all anchor tags worked
as intended, this commit removes the `rel=prefetch` that had been hardcoded on anchor tags in the
source code. Basically, we now prefetch all anchor tags by default so there is no need to manually
mark which tags need to be prefetched.